+++ component/SigninButton.tsx
@@ -1,0 +1,10 @@
+export default function SigninButton() {
+  return (
+    <button
+      type="submit"
+      className="mt-[30px] w-full px-5 py-4 bg-gradient-to-r from-[#6D6AFE] to-[#6AE3FE] rounded-md text-[#F5F5F5]"
+    >
+      로그인
+    </button>
+  );
+}
